requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
jwt==1.3.1
requests==2.31.0
